scraping/requirements.txt
boto3==1.18.21
pandas==1.3.4
python-dotenv==0.21.0
selenium==3.141.0
